SceneQueryRunner: decouple time range comparisons
Prior to this commit, `SceneQueryRunner` had special handling for
`SceneTimeRangeCompare` objects, explicitly searching for them in the
scene graph, adding additional queries, and transforming resulting
queries. This made it difficult to re-use the general behaviour of
'running additional queries' in other objects.

This commit introduces a new interface, `SceneRequestAdder`, which
can be implemented to inform the query runner that it should run
additional requests (returned by `getExtraRequests`) and transform the
results in some fashion.

Instead of searching the graph for `SceneTimeRangeCompare` objects,
the query runner searches for implementors of `SceneRequestAdder`
and uses those instead. The specifics of how it searches for these
is a little bit fuzzy and should probably be improved: it walks up the
graph until it finds at least one adder at the current level or in
any children of the current level; adds any others at that level
or in the children; then returns.

`SceneTimeRangeCompare` has been refactored to make use of this new
interface. I've also got a separate object which also implements
it which is working well including when both are enabled.