fix: docs builds in CI
Since the move to Turbo (done as part of adding scenes-react) docs builds
have been failing in CI ([example 1] and [example 2]). The config args
aren't being passed through to the website package since we're using turbo
now.

This commit adds the required extra '--' to the CI commands. That wasn't
enough though; because the turbo.json declares a dependency from any
'build' command to any others, running yarn docs:build resulted in
all the other packages' build commands running too, which would fail
because the --config flag would be passed to them and they don't understand
it. I've therefore also removed the dependency in turbo.json - I'm not sure
it was fully needed, but this might be a mistake? If so, we need another
solution.

I'm still not certain this will pass but will iterate until it does...